mutext guard op

+++ yolo/shortcut_layer.cpp
@@ -7,7 +7,7 @@ ShortcutLayer::ShortcutLayer(LayerPtr _from, LayerPtr prev)
 {
     output.copy_size(prev->get_output());
 
-    fprintf(stderr, "res                %4d x%4d x%4d   ->  %4d x%4d x%4d\n", from->get_output().nc(), from->get_output().nr(), from->get_output().k(), output.nc(), output.nr(), output.k());
+    fprintf(stderr, "res                %4d x%4d x%4d   ->  %4d x%4d x%4d\n", (int)from->get_output().nc(), (int)from->get_output().nr(), (int)from->get_output().k(), (int)output.nc(), (int)output.nr(), (int)output.k());
 }
 
 const tensor& ShortcutLayer::forward_layer(const tensor& input)
